Structure modules refactor
