+++ src/find-me-auth/find-me-auth.module.ts
@@ -1,13 +1,12 @@
 import { Module } from "@nestjs/common";
 import { JwtModule } from "@nestjs/jwt";
-import { FindMeSecurityModule } from "@src/modules/find-me-security/find-me-security.module";
-import { FindMeUsersModule } from "@src/modules/find-me-users/find-me-users.module";
-import { FindMeAuthService } from "./find-me-auth.service";
-import { FindMeAuthController } from "./find-me-auth.controller";
-import envConfig from "@src/config/env.config";
-import { JwtStrategy } from "@src/modules/find-me-auth/find-me-jwt.strategy";
+import { FindMeUsersModule } from "@src/find-me-users/find-me-users.module";
+import envConfig from "@src/find-me-commons/config/env.config";
+import { JwtStrategy } from "@src/find-me-auth/find-me-jwt.strategy";
 import { MongooseModule } from "@nestjs/mongoose";
-import { FindMeAuthToken, FindMeAuthTokenSchema } from "@src/modules/find-me-auth/schemas/find-me-auth-token.schema";
+import { FindMeAuthToken, FindMeAuthTokenSchema } from "@src/find-me-auth/schemas/find-me-auth-token.schema";
+import { FindMeAuthService } from "@src/find-me-auth/find-me-auth.service";
+import { FindMeAuthController } from "@src/find-me-auth/find-me-auth.controller";
 
 const secret = envConfig().encryptKey;
 
@@ -19,7 +18,6 @@ const secret = envConfig().encryptKey;
                 schema: FindMeAuthTokenSchema,
             },
         ]),
-        FindMeSecurityModule,
         FindMeUsersModule,
         JwtModule.register({ secret }),
     ],
