+++ src/find-me-commons/config/mongodb.config.ts
@@ -1,4 +1,4 @@
-import environmentConstants from "@src/constants/environment.constants";
+import environmentConstants from "@src/find-me-commons/constants/environment.constants";
 
 const dockerMongoUri = "mongodb://mongo/find-me-docker";
