+++ src/find-me-commons/dto/bad-request-exception.dto.ts
