+++ src/find-me-commons/dto/error-exception.dto.ts
