+++ src/find-me-users/find-me-users.module.ts
@@ -1,10 +1,9 @@
 import { Module } from "@nestjs/common";
 import { MongooseModule } from "@nestjs/mongoose";
-import { FindMeSecurityModule } from "@src/modules/find-me-security/find-me-security.module";
-import { FindMeUser, FindMeUserSchema } from "@src/modules/find-me-users/schemas/find-me-user.schema";
-import { FindMeUsersController } from "@src/modules/find-me-users/find-me-users.controller";
-import { FindMeUsersService } from "@src/modules/find-me-users/find-me-users.service";
-import { FindMeUserDeleteLog, FindMeUserDeleteLogSchema } from "@src/modules/find-me-users/schemas/find-me-user-delete-log.schema";
+import { FindMeUser, FindMeUserSchema } from "@src/find-me-users/schemas/find-me-user.schema";
+import { FindMeUsersController } from "@src/find-me-users/find-me-users.controller";
+import { FindMeUsersService } from "@src/find-me-users/find-me-users.service";
+import { FindMeUserDeleteLog, FindMeUserDeleteLogSchema } from "@src/find-me-users/schemas/find-me-user-delete-log.schema";
 
 @Module({
     imports: [
@@ -18,7 +17,6 @@ import { FindMeUserDeleteLog, FindMeUserDeleteLogSchema } from "@src/modules/fin
                 schema: FindMeUserDeleteLogSchema,
             },
         ]),
-        FindMeSecurityModule,
     ],
     controllers: [ FindMeUsersController ],
     providers: [ FindMeUsersService ],
